Refactor the test_orders.py to test_orders_class.py
Separated the test cases in test_orders.py to tests for the Orders class and the Orders related views
